+++ packages/docs/src/components/code.js
@@ -2,6 +2,7 @@
 import { jsx, Styled } from 'theme-ui'
 import Prism from '@theme-ui/prism'
 import { LiveProvider, LiveEditor, LivePreview, LiveError } from 'react-live'
+import * as components from '@theme-ui/components'
 
 const posts = [
   {
@@ -35,6 +36,15 @@ const posts = [
   },
 ]
 
+const images = {
+  nyc:
+    'https://images.unsplash.com/photo-1446776899648-aa78eefe8ed0?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=1080&fit=max&ixid=eyJhcHBfaWQiOjF9',
+  flatiron:
+    'https://images.unsplash.com/photo-1520222984843-df35ebc0f24d?ixlib=rb-1.2.1&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=1080&fit=max&ixid=eyJhcHBfaWQiOjF9',
+  logo:
+    'https://contrast.now.sh/fff/000?text=UI&size=96&fontSize=1.5&baseline=1',
+}
+
 const scope = {
   jsx,
   Styled,
@@ -44,9 +54,12 @@ const scope = {
     return <span {...props} sx={{ cursor: 'pointer' }} />
   },
   posts,
+  images,
+  // todo: Link conflict
+  ...components,
 }
 
-const transformCode = src => `/** @jsx jsx */\n${src}`
+const transformCode = src => `/** @jsx jsx */\n<>${src}</>`
 
 const liveTheme = { styles: [] }
 
@@ -88,7 +101,8 @@ export const LiveCode = ({ children, preview, xray }) => {
       </div>
       <Styled.pre
         sx={{
-          my: 0,
+          mt: 0,
+          mb: 3,
         }}>
         <LiveEditor padding={0} />
       </Styled.pre>