+++ packages/editor/src/SxMargin.js
@@ -1,0 +1,11 @@
+/** @jsx jsx */
+import { jsx } from 'theme-ui'
+import SxSpace from './SxSpace'
+
+export const SxMargin = props =>
+  <SxSpace
+    {...props}
+    property='margin'
+  />
+
+export default SxMargin