+++ packages/editor/src/SxPadding.js
@@ -1,0 +1,11 @@
+/** @jsx jsx */
+import { jsx } from 'theme-ui'
+import SxSpace from './SxSpace'
+
+export const SxPadding = props =>
+  <SxSpace
+    {...props}
+    property='padding'
+  />
+
+export default SxPadding