+++ packages/editor/src/index.js
@@ -14,7 +14,15 @@ export { default as Colors } from './Colors'
 export { default as Modes } from './Modes'
 export { default as Typography } from './Typography'
 
-export {
-  EditorContext,
-  useEditor,
-} from './context'
+// todo: figure out name
+export { default as StylesForm } from './StylesForm'
+export { default as ThemeColorPicker } from './ThemeColorPicker'
+
+// sx style object forms
+export { default as SxColors } from './SxColors'
+export { default as SxSpace } from './SxSpace'
+export { default as SxMargin } from './SxMargin'
+export { default as SxPadding } from './SxPadding'
+export { default as SxTypography } from './SxTypography'
+
+export { EditorContext, useEditor } from './context'