+++ packages/theme-ui/src/color-modes.js
@@ -5,10 +5,11 @@ import { useThemeUI } from './context'
 import { createColorStyles } from './custom-properties'
 
 const STORAGE_KEY = 'theme-ui-color-mode'
+const HAS_STORAGE = typeof Storage !== 'undefined'
 
 const storage = {
-  get: init => window.localStorage.getItem(STORAGE_KEY) || init,
-  set: value => window.localStorage.setItem(STORAGE_KEY, value),
+  get: init => HAS_STORAGE && window.localStorage.getItem(STORAGE_KEY) || init,
+  set: value => HAS_STORAGE && window.localStorage.setItem(STORAGE_KEY, value),
 }
 
 export const getMediaQuery = () => {
@@ -18,7 +19,8 @@ export const getMediaQuery = () => {
   return dark && mql.matches
 }
 
-const getName = (theme) => theme.initialColorModeName || theme.initialColorMode || 'default'
+const getName = theme =>
+  theme.initialColorModeName || theme.initialColorMode || 'default'
 
 export const useColorState = theme => {
   const [mode, setMode] = useState(getName(theme))
@@ -28,7 +30,8 @@ export const useColorState = theme => {
     const stored = storage.get()
     document.body.classList.remove('theme-ui-' + stored)
     const dark = getMediaQuery()
-    if (!stored && dark && theme.useColorSchemeMediaQuery) return setMode('dark')
+    if (!stored && dark && theme.useColorSchemeMediaQuery)
+      return setMode('dark')
     if (!stored || stored === mode) return
     setMode(stored)
   }, [])
@@ -46,7 +49,7 @@ export const useColorState = theme => {
     ) {
       console.warn(
         'The `initialColorMode` value should be a unique name' +
-          'and cannot reference a key in `theme.colors.modes`.'
+          ' and cannot reference a key in `theme.colors.modes`.'
       )
     }
   }
@@ -54,7 +57,7 @@ export const useColorState = theme => {
   return [mode, setMode]
 }
 
-export const useColorMode = initialMode => {
+export const useColorMode = () => {
   const { colorMode, setColorMode } = useThemeUI()
 
   if (typeof setColorMode !== 'function') {